Bootstrapped dataframe with augmented predictions from each sample.

Bootstrapped dataframe with augmented predictions from each sample.

boot_augment(
  data,
  times = 25,
  strata = NULL,
  apparent = FALSE,
  ..f,
  ...,
  augment.args = list()
)

Arguments

data
Dataframe (or tibble) from which variables are to be taken.

times
The number of bootstrap samples.

strata
A variable that is used to conduct stratified sampling. When not NULL, each bootstrap sample is created within the stratification variable. This could be a single character value or a variable name that corresponds to a variable that exists in the data frame.

apparent
A logical. Should an extra resample be added where the analysis and holdout subset are the entire data set. This is required for some estimators used by the summary function that require the apparent error rate.

..f
A function, or function name as a string.

...
<dynamic> Arguments for .fn.

augment.args
A list of arguments to be used in the relevant S3 method.

Value

A tibble::tibble() with information about data points.

Note

For available methods, see- https://indrajeetpatil.github.io/broomExtra/articles/available_methods.html

Methods

See the following help topics for more details about individual methods:broom

augment.betareg: betareg

augment.clm: clm

augment.coxph: coxph

augment.decomposed.ts: decomposed.ts

augment.drc: drc

augment.factanal: factanal

augment.felm: felm

augment.glm: glm

augment.glmRob: glmRob

augment.htest: htest

augment.ivreg: ivreg

augment.kmeans: kmeans

augment.lm: lm

augment.lmRob: lmRob

augment.loess: loess

augment.Mclust: Mclust

augment.mjoint: mjoint

augment.nlrq: nlrq

augment.nls: nls

augment.pam: pam

augment.plm: plm

augment.poLCA: poLCA

augment.polr: polr

augment.prcomp: prcomp

augment.rlm: rlm

augment.rma: rma

augment.robustbase.glmrob: glmrob

augment.robustbase.lmrob: lmrob

augment.rq: rq

augment.rqs: rqs

augment.smooth.spline: smooth.spline

augment.speedlm: speedlm

augment.stl: stl

augment.survreg: survreg

null_tidiers: NULL

broom.mixed

augment.ranef.mer: ranef.mer

brms_tidiers: brmsfit

glmmadmb_tidiers: glmmadmb

glmmTMB_tidiers: glmmTMB

lme4_tidiers: merMod

nlme_tidiers: gls, lme

See also

Examples

set.seed(123)

# example-1: linear model
broomExtra::boot_augment(
  data = mtcars,
  times = 10,
  ..f = stats::lm,
  formula = mpg ~ wt,
  na.action = na.omit
)

#> # A tibble: 320 x 11
#>    splits          id          .rownames            mpg    wt .fitted  .resid
#>    <list>          <chr>       <chr>              <dbl> <dbl>   <dbl>   <dbl>
#>  1 <split [32/11]> Bootstrap01 Maserati Bora       15    3.57   18.0   2.96  
#>  2 <split [32/11]> Bootstrap01 Cadillac Fleetwood  10.4  5.25    7.71 -2.69  
#>  3 <split [32/11]> Bootstrap01 Honda Civic         30.4  1.62   29.9  -0.522 
#>  4 <split [32/11]> Bootstrap01 Merc 450SLC         15.2  3.78   16.7   1.48  
#>  5 <split [32/11]> Bootstrap01 Datsun 710          22.8  2.32   25.6   2.78  
#>  6 <split [32/11]> Bootstrap01 Merc 280            19.2  3.44   18.8  -0.449 
#>  7 <split [32/11]> Bootstrap01 Fiat 128            32.4  2.2    26.3  -6.09  
#>  8 <split [32/11]> Bootstrap01 Dodge Challenger    15.5  3.52   18.3   2.76  
#>  9 <split [32/11]> Bootstrap01 Merc 280C           17.8  3.44   18.8   0.951 
#> 10 <split [32/11]> Bootstrap01 Hornet Sportabout   18.7  3.44   18.8   0.0506
#>    .std.resid   .hat .sigma    .cooksd
#>         <dbl>  <dbl>  <dbl>      <dbl>
#>  1    -1.12   0.0458   2.69 0.0301    
#>  2     1.15   0.250    2.69 0.218     
#>  3     0.205  0.111    2.75 0.00261   
#>  4    -0.563  0.0582   2.74 0.00978   
#>  5    -1.05   0.0500   2.70 0.0293    
#>  6     0.170  0.0401   2.75 0.000600  
#>  7     2.32   0.0574   2.49 0.164     
#>  8    -1.04   0.0435   2.70 0.0248    
#>  9    -0.359  0.0401   2.75 0.00268   
#> 10    -0.0191 0.0401   2.75 0.00000761
#> # ... with 310 more rows

# example-2: linear mixed-effects model
library(lme4)

broomExtra::boot_augment(
  data = sleepstudy,
  times = 25,
  ..f = lme4::lmer,
  formula = Reaction ~ Days + (Days | Subject)
)

#> Warning: Model failed to converge with max|grad| = 0.00416738 (tol = 0.002, component 1)

#> Warning: Model failed to converge with max|grad| = 0.344884 (tol = 0.002, component 1)

#> Warning: Model failed to converge with max|grad| = 0.00307876 (tol = 0.002, component 1)

#> # A tibble: 4,500 x 17
#>    splits           id          .rownames Reaction  Days Subject .fitted .resid
#>    <list>           <chr>       <chr>        <dbl> <dbl> <fct>     <dbl>  <dbl>
#>  1 <split [180/69]> Bootstrap01 145           315.     4 369        305.   9.37
#>  2 <split [180/69]> Bootstrap01 165           279.     4 371        293. -13.5 
#>  3 <split [180/69]> Bootstrap01 57            454.     6 332        341. 114.  
#>  4 <split [180/69]> Bootstrap01 148           348.     7 369        336.  12.2 
#>  5 <split [180/69]> Bootstrap01 163           278.     2 371        269.   8.61
#>  6 <split [180/69]> Bootstrap01 161           270.     0 371        246.  24.0 
#>  7 <split [180/69]> Bootstrap01 66            338.     5 333        324.  14.0 
#>  8 <split [180/69]> Bootstrap01 4             321.     3 308        309.  12.4 
#>  9 <split [180/69]> Bootstrap01 74            255.     3 334        262.  -7.52
#> 10 <split [180/69]> Bootstrap01 133           327.     2 352        323.   3.46
#>      .hat .cooksd .fixed   .mu .offset .sqrtXwt .sqrtrwt .weights .wtres
#>     <dbl>   <dbl>  <dbl> <dbl>   <dbl>    <dbl>    <dbl>    <dbl>  <dbl>
#>  1 0.0689 0.00686   297.  305.       0        1        1        1   9.37
#>  2 0.0689 0.0143    297.  293.       0        1        1        1 -13.5 
#>  3 0.164  2.97      319.  341.       0        1        1        1 114.  
#>  4 0.157  0.0321    330.  336.       0        1        1        1  12.2 
#>  5 0.124  0.0117    275.  269.       0        1        1        1   8.61
#>  6 0.223  0.208     253.  246.       0        1        1        1  24.0 
#>  7 0.109  0.0265    308.  324.       0        1        1        1  14.0 
#>  8 0.155  0.0331    286.  309.       0        1        1        1  12.4 
#>  9 0.134  0.00994   286.  262.       0        1        1        1  -7.52
#> 10 0.213  0.00404   275.  323.       0        1        1        1   3.46
#> # ... with 4,490 more rows