Bootstrapped dataframe with model summaries from each sample.

Bootstrapped dataframe with model summaries from each sample.

boot_glance(data, times = 25, strata = NULL, apparent = FALSE, ..f, ...)

Arguments

data
Dataframe (or tibble) from which variables are to be taken.

times
The number of bootstrap samples.

strata
A variable that is used to conduct stratified sampling. When not NULL, each bootstrap sample is created within the stratification variable. This could be a single character value or a variable name that corresponds to a variable that exists in the data frame.

apparent
A logical. Should an extra resample be added where the analysis and holdout subset are the entire data set. This is required for some estimators used by the summary function that require the apparent error rate.

..f
A function, or function name as a string.

...
Arguments to function.

These dots support tidy-dots features.

Note

For available methods, see- https://indrajeetpatil.github.io/broomExtra/articles/available_methods.html

Methods

See the following help topics for more details about individual methods:broom

durbinWatsonTest_tidiers: durbinWatsonTest

glance.aareg: aareg

glance.aov: aov

glance.Arima: Arima

glance.betareg: betareg

glance.biglm: biglm

glance.binDesign: binDesign

glance.cch: cch

glance.clm: clm

glance.clmm: clmm

glance.coxph: coxph

glance.cv.glmnet: cv.glmnet

glance.drc: drc

glance.ergm: ergm

glance.factanal: factanal

glance.felm: felm

glance.fitdistr: fitdistr

glance.Gam: Gam

glance.garch: garch

glance.geeglm: geeglm

glance.glm: glm

glance.glmnet: glmnet

glance.glmRob: glmRob

glance.gmm: gmm

glance.ivreg: ivreg

glance.kmeans: kmeans

glance.lavaan: lavaan

glance.lm: lm

glance.lmodel2: lmodel2

glance.lmRob: lmRob

glance.Mclust: Mclust

glance.mjoint: mjoint

glance.muhaz: muhaz

glance.multinom: multinom

glance.nlrq: nlrq

glance.nls: nls

glance.orcutt: orcutt

glance.pam: pam

glance.plm: plm

glance.poLCA: poLCA

glance.polr: polr

glance.pyears: pyears

glance.ridgelm: ridgelm

glance.rlm: rlm

glance.rma: rma

glance.robustbase.lmrob: lmrob

glance.rq: rq

glance.smooth.spline: smooth.spline

glance.speedglm: speedglm

glance.speedlm: speedlm

glance.survdiff: survdiff

glance.survexp: survexp

glance.survfit: survfit

glance.survreg: survreg

glance.svyglm: svyglm

glance.svyolr: svyolr

list_tidiers: list

mgcv_glance_gam: gam

null_tidiers: NULL

tidy.htest: htest

broom.mixed

brms_tidiers: brmsfit

glmmadmb_tidiers: glmmadmb

glmmTMB_tidiers: glmmTMB

lme4_tidiers: merMod

nlme_tidiers: lme

rstanarm_tidiers: stanreg

See also

Examples

set.seed(123)

# example-1: linear model
broomExtra::boot_glance(
  data = mtcars,
  times = 500,
  ..f = stats::lm,
  formula = mpg ~ wt,
  na.action = na.omit
)

#> # A tibble: 500 x 14
#>    splits id    r.squared adj.r.squared sigma statistic  p.value    df logLik
#>    <list> <chr>     <dbl>         <dbl> <dbl>     <dbl>    <dbl> <dbl>  <dbl>
#>  1 <spli~ Boot~     0.799         0.792  2.71     119.  5.63e-12     1  -76.2
#>  2 <spli~ Boot~     0.726         0.717  2.49      79.6 6.06e-10     1  -73.6
#>  3 <spli~ Boot~     0.754         0.746  3.29      92.1 1.18e-10     1  -82.4
#>  4 <spli~ Boot~     0.731         0.722  2.74      81.7 4.59e-10     1  -76.6
#>  5 <spli~ Boot~     0.700         0.690  3.33      69.9 2.49e- 9     1  -82.9
#>  6 <spli~ Boot~     0.830         0.824  2.19     146.  4.70e-13     1  -69.4
#>  7 <spli~ Boot~     0.768         0.761  3.61      99.5 4.87e-11     1  -85.4
#>  8 <spli~ Boot~     0.550         0.535  3.01      36.6 1.20e- 6     1  -79.7
#>  9 <spli~ Boot~     0.815         0.809  2.66     132.  1.57e-12     1  -75.7
#> 10 <spli~ Boot~     0.680         0.670  3.32      63.8 6.43e- 9     1  -82.8
#> # ... with 490 more rows, and 5 more variables: AIC <dbl>, BIC <dbl>,
#> #   deviance <dbl>, df.residual <int>, nobs <int>

# example-2: linear mixed-effects model
library(lme4)

broomExtra::boot_glance(
  data = sleepstudy,
  times = 25,
  ..f = lme4::lmer,
  formula = Reaction ~ Days + (Days | Subject)
)

#> Warning: Model failed to converge with max|grad| = 0.00205594 (tol = 0.002, component 1)

#> Warning: Model failed to converge with max|grad| = 0.00239836 (tol = 0.002, component 1)

#> Warning: Model failed to converge with max|grad| = 0.00226858 (tol = 0.002, component 1)

#> Warning: Model failed to converge with max|grad| = 0.00237924 (tol = 0.002, component 1)

#> Warning: Model failed to converge with max|grad| = 0.00278758 (tol = 0.002, component 1)

#> Warning: Model failed to converge with max|grad| = 0.00341433 (tol = 0.002, component 1)

#> Warning: Model failed to converge with max|grad| = 0.00943341 (tol = 0.002, component 1)

#> Warning: Model failed to converge with max|grad| = 0.00522705 (tol = 0.002, component 1)

#> # A tibble: 25 x 8
#>    splits           id          sigma logLik   AIC   BIC REMLcrit df.residual
#>    <list>           <chr>       <dbl>  <dbl> <dbl> <dbl>    <dbl>       <int>
#>  1 <split [180/70]> Bootstrap01  27.1  -879. 1769. 1788.    1757.         174
#>  2 <split [180/71]> Bootstrap02  20.1  -838. 1687. 1706.    1675.         174
#>  3 <split [180/71]> Bootstrap03  28.9  -889. 1790. 1809.    1778.         174
#>  4 <split [180/58]> Bootstrap04  25.9  -875. 1762. 1781.    1750.         174
#>  5 <split [180/68]> Bootstrap05  22.1  -850. 1712. 1731.    1700.         174
#>  6 <split [180/68]> Bootstrap06  21.2  -847. 1705. 1724.    1693.         174
#>  7 <split [180/60]> Bootstrap07  22.2  -850. 1713. 1732.    1701.         174
#>  8 <split [180/67]> Bootstrap08  16.9  -808. 1627. 1646.    1615.         174
#>  9 <split [180/69]> Bootstrap09  20.6  -843. 1697. 1716.    1685.         174
#> 10 <split [180/68]> Bootstrap10  23.7  -859. 1730. 1749.    1718.         174
#> # ... with 15 more rows