Bootstrapped dataframe with estimates from each sample.

Bootstrapped dataframe with estimates from each sample.

boot_tidy(data, times = 25, strata = NULL, apparent = FALSE, ..f,
  ..., tidy.args = list())

Arguments

data
Dataframe (or tibble) from which variables are to be taken.

times
The number of bootstrap samples.

strata
A variable that is used to conduct stratified sampling. When not NULL, each bootstrap sample is created within the stratification variable.

apparent
A logical. Should an extra resample be added where the analysis and holdout subset are the entire data set. This is required for some estimators used by the summary function that require the apparent error rate.

..f
A function, or function name as a string.

...
Arguments to function.

These dots support tidy-dots features.

tidy.args
A list of arguments to be used in the relevant S3 method.

Value

A tibble::tibble() with information about model components.

Note

For available methods, see- https://indrajeetpatil.github.io/broomExtra/articles/available_methods.html

Methods

See the following help topics for more details about individual methods: broom

durbinWatsonTest_tidiers: durbinWatsonTest

list_tidiers: list

metafor_tidiers: rma

mgcv_tidy_gam: gam

null_tidiers: NULL

sp_tidiers: Line, Lines, Polygon, Polygons, SpatialLinesDataFrame, SpatialPolygons, SpatialPolygonsDataFrame

tidy.aareg: aareg

tidy.acf: acf

tidy.anova: anova

tidy.aov: aov

tidy.aovlist: aovlist

tidy.Arima: Arima

tidy.betareg: betareg

tidy.biglm: biglm

tidy.binDesign: binDesign

tidy.binWidth: binWidth

tidy.boot: boot

tidy.btergm: btergm

tidy.cch: cch

tidy.cld: cld

tidy.clm: clm

tidy.clmm: clmm

tidy.coeftest: coeftest

tidy.confint.glht: confint.glht

tidy.confusionMatrix: confusionMatrix

tidy.coxph: coxph

tidy.cv.glmnet: cv.glmnet

tidy.density: density

tidy.dist: dist

tidy.drc: drc

tidy.emmGrid: emmGrid

tidy.epi.2by2: epi.2by2

tidy.ergm: ergm

tidy.factanal: factanal

tidy.felm: felm

tidy.fitdistr: fitdistr

tidy.ftable: ftable

tidy.Gam: Gam

tidy.garch: garch

tidy.geeglm: geeglm

tidy.glht: glht

tidy.glm: glm

tidy.glmnet: glmnet

tidy.glmRob: glmRob

tidy.gmm: gmm

tidy.htest: htest

tidy.ivreg: ivreg

tidy.kappa: kappa

tidy.kde: kde

tidy.Kendall: Kendall

tidy.kmeans: kmeans

tidy.lavaan: lavaan

tidy.lm.beta: lm.beta, summary.lm.beta

tidy.lm: lm, summary.lm

tidy.lmodel2: lmodel2

tidy.lmRob: lmRob

tidy.lsmobj: lsmobj

tidy.manova: manova

tidy.map: map

tidy.Mclust: Mclust

tidy.mediate: mediate

tidy.mjoint: mjoint

tidy.mle2: mle2

tidy.mlm: mlm

tidy.muhaz: muhaz

tidy.multinom: multinom

tidy.nlrq: nlrq

tidy.nls: nls

tidy.orcutt: orcutt

tidy.pairwise.htest: pairwise.htest

tidy.pam: pam

tidy.plm: plm

tidy.poLCA: poLCA

tidy.polr: polr

tidy.power.htest: power.htest

tidy.prcomp: prcomp

tidy.pyears: pyears

tidy.rcorr: rcorr

tidy.ref.grid: ref.grid

tidy.regsubsets: regsubsets

tidy.ridgelm: ridgelm

tidy.rlm: rlm

tidy.roc: roc

tidy.rq: rq

tidy.rqs: rqs

tidy.spec: spec

tidy.speedlm: speedlm

tidy.summary.glht: summary.glht

tidy.survdiff: survdiff

tidy.survexp: survexp

tidy.survfit: survfit

tidy.survreg: survreg

tidy.svyglm: svyglm

tidy.svyolr: svyolr

tidy.systemfit: systemfit

tidy.table: table

tidy.ts: ts

tidy.TukeyHSD: TukeyHSD

tidy.zoo: zoo

broom.mixed

brms_tidiers: brmsfit

gamlss_tidiers: gamlss

glmmadmb_tidiers: glmmadmb

glmmTMB_tidiers: glmmTMB

lme4_tidiers: merMod, rlmerMod

mcmc_tidiers: mcmc, mcmc.list, MCMCglmm, rjags, stanfit

nlme_tidiers: gls, lme

rstanarm_tidiers: stanreg

tidy.TMB: TMB

rsample

tidy.rsplit: nested_cv, rset, rsplit, vfold_cv

See also

Examples

set.seed(123)

# example-1: linear model
broomExtra::boot_tidy(
  data = mtcars,
  times = 500,
  ..f = stats::lm,
  formula = mpg ~ wt,
  na.action = na.omit,
  tidy.args = list(conf.int = TRUE, conf.level = 0.50)
)

#> # A tibble: 1,000 x 8
#>    id         term      estimate std.error statistic  p.value conf.low conf.high
#>    <chr>      <chr>        <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl>
#>  1 Bootstrap~ (Interce~    39.7      1.73      22.9  1.44e-20    38.5      40.9 
#>  2 Bootstrap~ wt           -6.10     0.558    -10.9  5.63e-12    -6.48     -5.72
#>  3 Bootstrap~ (Interce~    35.0      1.79      19.5  1.29e-18    33.8      36.2 
#>  4 Bootstrap~ wt           -4.76     0.533     -8.92 6.06e-10    -5.12     -4.39
#>  5 Bootstrap~ (Interce~    37.2      1.93      19.2  2.04e-18    35.9      38.5 
#>  6 Bootstrap~ wt           -5.11     0.532     -9.60 1.18e-10    -5.47     -4.74
#>  7 Bootstrap~ (Interce~    36.2      1.98      18.3  7.90e-18    34.8      37.5 
#>  8 Bootstrap~ wt           -5.24     0.580     -9.04 4.59e-10    -5.64     -4.85
#>  9 Bootstrap~ (Interce~    36.5      2.18      16.8  8.60e-17    35.0      38.0 
#> 10 Bootstrap~ wt           -5.12     0.613     -8.36 2.49e- 9    -5.54     -4.71
#> # ... with 990 more rows

# example-2: linear mixed-effects model
library(lme4)

broomExtra::boot_tidy(
  data = sleepstudy,
  times = 25,
  ..f = lme4::lmer,
  formula = Reaction ~ Days + (Days | Subject),
  tidy.args = list(effects = "fixed")
)

#> Warning: Model failed to converge with max|grad| = 0.00205594 (tol = 0.002, component 1)

#> Warning: Model failed to converge with max|grad| = 0.00239836 (tol = 0.002, component 1)

#> Warning: Model failed to converge with max|grad| = 0.00226858 (tol = 0.002, component 1)

#> Warning: Model failed to converge with max|grad| = 0.00237924 (tol = 0.002, component 1)

#> Warning: Model failed to converge with max|grad| = 0.00278758 (tol = 0.002, component 1)

#> Warning: Model failed to converge with max|grad| = 0.00341433 (tol = 0.002, component 1)

#> Warning: Model failed to converge with max|grad| = 0.00943341 (tol = 0.002, component 1)

#> Warning: Model failed to converge with max|grad| = 0.00522705 (tol = 0.002, component 1)

#> # A tibble: 50 x 6
#>    id          effect term        estimate std.error statistic
#>    <chr>       <chr>  <chr>          <dbl>     <dbl>     <dbl>
#>  1 Bootstrap01 fixed  (Intercept)   248.        7.94     31.3 
#>  2 Bootstrap01 fixed  Days           10.8       1.44      7.49
#>  3 Bootstrap02 fixed  (Intercept)   255.        8.42     30.3 
#>  4 Bootstrap02 fixed  Days            9.68      1.73      5.59
#>  5 Bootstrap03 fixed  (Intercept)   245.        7.56     32.5 
#>  6 Bootstrap03 fixed  Days           11.3       1.47      7.65
#>  7 Bootstrap04 fixed  (Intercept)   254.        6.85     37.1 
#>  8 Bootstrap04 fixed  Days           10.2       1.77      5.78
#>  9 Bootstrap05 fixed  (Intercept)   250.        8.19     30.5 
#> 10 Bootstrap05 fixed  Days            9.78      1.66      5.88
#> # ... with 40 more rows