Model performance summary dataframes using broom and easystats.

Computes indices of model performance for regression models.

glance_performance(x, ...)

Arguments

x
model or other R object to convert to single-row data frame

...
other arguments passed to methods

Value

A data frame (with one row) and one column per "index".

Details

The function will attempt to get these details either using broom::glance or performance::model_performance. If both function provide model performance measure summaries, the function will try to combine them into a single dataframe.

Examples

set.seed(123)
mod <- lm(mpg ~ wt + cyl, data = mtcars)
broomExtra::glance_performance(mod)

#> # A tibble: 1 x 13
#>   r.squared adj.r.squared sigma statistic  p.value    df loglik   aic   bic
#>       <dbl>         <dbl> <dbl>     <dbl>    <dbl> <dbl>  <dbl> <dbl> <dbl>
#> 1     0.830         0.819  2.57      70.9 6.81e-12     2  -74.0  156.  162.
#>   deviance df.residual  nobs  rmse
#>      <dbl>       <int> <int> <dbl>
#> 1     191.          29    32  2.44