Augmented data from grouped analysis of any function that has data argument in its function call.

Augmented data from grouped analysis of any function that has data argument in its function call.

grouped_augment(data, grouping.vars, ..f, ..., augment.args = list())

Arguments

data
Dataframe (or tibble) from which variables are to be taken.

grouping.vars
Grouping variables.

..f
A function, or function name as a string.

...
Arguments to function.

These dots support tidy-dots features.

augment.args
A list of arguments to be used in the relevant S3 method.

Examples

set.seed(123)
# to speed up computation, let's use only 50% of the data

# linear model
broomExtra::grouped_augment(
  data = dplyr::sample_frac(tbl = ggplot2::diamonds, size = 0.5),
  grouping.vars = c(cut, color),
  formula = price ~ carat - 1,
  ..f = stats::lm,
  na.action = na.omit
)

#> # A tibble: 26,970 x 10
#>    cut   color price carat .fitted  .resid .std.resid    .hat .sigma   .cooksd
#>    <ord> <ord> <int> <dbl>   <dbl>   <dbl>      <dbl>   <dbl>  <dbl>     <dbl>
#>  1 Fair  D      1323  0.5    2581.  1258.     -0.658  0.00278  1920. 0.00121  
#>  2 Fair  D      3855  0.91   4698.   843.     -0.442  0.00922  1923. 0.00182  
#>  3 Fair  D      5096  1      5162.    66.2    -0.0348 0.0111   1926. 0.0000136
#>  4 Fair  D       997  0.31   1600.   603.     -0.315  0.00107  1924. 0.000107 
#>  5 Fair  D      1895  0.7    3614.  1719.     -0.901  0.00546  1916. 0.00445  
#>  6 Fair  D      3205  0.9    4646.  1441.     -0.757  0.00902  1919. 0.00521  
#>  7 Fair  D      1770  0.7    3614.  1844.     -0.966  0.00546  1914. 0.00512  
#>  8 Fair  D      9704  1.52   7847. -1857.      0.984  0.0257   1914. 0.0256   
#>  9 Fair  D      4441  1.03   5317.   876.     -0.461  0.0118   1923. 0.00254  
#> 10 Fair  D      5078  0.93   4801.  -277.      0.146  0.00963  1925. 0.000206 
#> # ... with 26,960 more rows

# linear mixed effects model
broomExtra::grouped_augment(
  data = dplyr::sample_frac(tbl = ggplot2::diamonds, size = 0.5),
  grouping.vars = cut,
  ..f = lme4::lmer,
  formula = price ~ carat + (carat | color) - 1,
  control = lme4::lmerControl(optimizer = "bobyqa")
)

#> # A tibble: 26,970 x 15
#>    cut   price carat color .fitted  .resid    .hat .cooksd .fixed     .mu
#>    <ord> <int> <dbl> <ord>   <dbl>   <dbl>   <dbl>   <dbl>  <dbl>   <dbl>
#>  1 Fair   2117  0.7  I       1851.  2.66e2 0.0148  4.54e-4  1948.   1851.
#>  2 Fair   7294  2.01 F      10416. -3.12e3 0.0515  2.35e-1  5592.  10416.
#>  3 Fair    579  0.31 E       -386.  9.65e2 0.0226  9.26e-3   862.   -386.
#>  4 Fair   7211  1.5  H       7047.  1.64e2 0.00908 1.04e-4  4173.   7047.
#>  5 Fair    497  0.34 J        276.  2.21e2 0.0318  6.95e-4   946.    276.
#>  6 Fair    727  0.5  G        989. -2.62e2 0.0125  3.70e-4  1391.    989.
#>  7 Fair   2372  0.71 D       2686. -3.14e2 0.0115  4.87e-4  1975.   2686.
#>  8 Fair   5846  1.5  J       5576.  2.70e2 0.0164  5.18e-4  4173.   5576.
#>  9 Fair   2782  0.84 G       3026. -2.44e2 0.00658 1.67e-4  2337.   3026.
#> 10 Fair   1848  0.7  I       1851. -3.28e0 0.0148  6.92e-8  1948.   1851.
#> # ... with 26,960 more rows, and 5 more variables: .offset <dbl>,
#> #   .sqrtXwt <dbl>, .sqrtrwt <dbl>, .weights <dbl>, .wtres <dbl>