Model summary output from grouped analysis of any function that has data argument in its function call.

Model summary output from grouped analysis of any function that has data argument in its function call.

grouped_glance(data, grouping.vars, ..f, ...)

Arguments

data
Dataframe (or tibble) from which variables are to be taken.

grouping.vars
Grouping variables.

..f
A function, or function name as a string.

...
Arguments to function.

These dots support tidy-dots features.

Examples

set.seed(123)
# to speed up computation, let's use only 50% of the data

# linear model
broomExtra::grouped_glance(
  data = dplyr::sample_frac(tbl = ggplot2::diamonds, size = 0.5),
  grouping.vars = c(cut, color),
  formula = price ~ carat - 1,
  ..f = stats::lm,
  na.action = na.omit
)

#> # A tibble: 35 x 13
#>    cut   color r.squared adj.r.squared sigma statistic   p.value    df logLik
#>    <ord> <ord>     <dbl>         <dbl> <dbl>     <dbl>     <dbl> <int>  <dbl>
#>  1 Fair  D         0.893         0.892 1913.      654. 1.16e- 39     1  -709.
#>  2 Fair  E         0.877         0.876 1761.      786. 6.59e- 52     1  -987.
#>  3 Fair  F         0.873         0.872 1888.     1084. 1.20e- 72     1 -1424.
#>  4 Fair  G         0.857         0.857 1976.     1011. 5.70e- 73     1 -1522.
#>  5 Fair  H         0.885         0.884 2065.     1108. 1.66e- 69     1 -1312.
#>  6 Fair  I         0.886         0.885 2010.      724. 1.14e- 45     1  -848.
#>  7 Fair  J         0.910         0.909 1996.      689. 2.59e- 37     1  -622.
#>  8 Good  D         0.848         0.847 1827.     1817. 2.47e-135     1 -2919.
#>  9 Good  E         0.869         0.869 1636.     3365. 1.00e-225     1 -4471.
#> 10 Good  F         0.873         0.872 1700.     3054. 1.24e-201     1 -3959.
#> # ... with 25 more rows, and 4 more variables: AIC <dbl>, BIC <dbl>,
#> #   deviance <dbl>, df.residual <int>

# linear mixed effects model
broomExtra::grouped_glance(
  data = dplyr::sample_frac(tbl = ggplot2::diamonds, size = 0.5),
  grouping.vars = cut,
  ..f = lme4::lmer,
  formula = price ~ carat + (carat | color) - 1,
  control = lme4::lmerControl(optimizer = "bobyqa")
)

#> # A tibble: 5 x 7
#>   cut       sigma  logLik     AIC     BIC REMLcrit df.residual
#>   <ord>     <dbl>   <dbl>   <dbl>   <dbl>    <dbl>       <int>
#> 1 Fair      1541.  -6945.  13901.  13924.   13891.         786
#> 2 Good      1361. -21622.  43254.  43283.   43244.        2496
#> 3 Very Good 1372. -52554. 105117. 105151.  105107.        6072
#> 4 Premium   1574. -60379. 120767. 120801.  120757.        6868
#> 5 Ideal     1294. -92129. 184268. 184304.  184258.       10723