Tidy output from grouped analysis of any function that has data argument in its function call.

Tidy output from grouped analysis of any function that has data argument in its function call.

grouped_tidy(data, grouping.vars, ..f, ..., tidy.args = list())

Arguments

data
Dataframe (or tibble) from which variables are to be taken.

grouping.vars
Grouping variables.

..f
A function, or function name as a string.

...
Arguments to function.

These dots support tidy-dots features.

tidy.args
A list of arguments to be used in the relevant S3 method.

Examples

set.seed(123)
# to speed up computation, let's use only 50% of the data

# linear model
broomExtra::grouped_tidy(
  data = dplyr::sample_frac(tbl = ggplot2::diamonds, size = 0.5),
  grouping.vars = c(cut, color),
  formula = price ~ carat - 1,
  ..f = stats::lm,
  na.action = na.omit,
  tidy.args = list(quick = TRUE)
)

#> # A tibble: 35 x 4
#>    cut   color term  estimate
#>    <ord> <ord> <chr>    <dbl>
#>  1 Fair  D     carat    5162.
#>  2 Fair  E     carat    5020.
#>  3 Fair  F     carat    4780.
#>  4 Fair  G     carat    4370.
#>  5 Fair  H     carat    4329.
#>  6 Fair  I     carat    4196.
#>  7 Fair  J     carat    3934.
#>  8 Good  D     carat    5213.
#>  9 Good  E     carat    5114.
#> 10 Good  F     carat    5165.
#> # ... with 25 more rows

# linear mixed effects model
broomExtra::grouped_tidy(
  data = dplyr::sample_frac(tbl = ggplot2::diamonds, size = 0.5),
  grouping.vars = cut,
  ..f = lme4::lmer,
  formula = price ~ carat + (carat | color) - 1,
  control = lme4::lmerControl(optimizer = "bobyqa"),
  tidy.args = list(conf.int = TRUE, conf.level = 0.99)
)

#> # A tibble: 25 x 9
#>    cut   effect  group  term     estimate std.error statistic conf.low conf.high
#>    <ord> <chr>   <chr>  <chr>       <dbl>     <dbl>     <dbl>    <dbl>     <dbl>
#>  1 Fair  fixed   NA     carat     2.78e+3      228.      12.2    2194.     3370.
#>  2 Fair  ran_pa~ color  sd__(In~  2.32e+3       NA       NA        NA        NA 
#>  3 Fair  ran_pa~ color  sd__car~  3.75e+3       NA       NA        NA        NA 
#>  4 Fair  ran_pa~ color  cor__(I~ -9.91e-1       NA       NA        NA        NA 
#>  5 Fair  ran_pa~ Resid~ sd__Obs~  1.54e+3       NA       NA        NA        NA 
#>  6 Good  fixed   NA     carat     5.70e+3      214.      26.6    5146.     6249.
#>  7 Good  ran_pa~ color  sd__(In~  2.61e+3       NA       NA        NA        NA 
#>  8 Good  ran_pa~ color  sd__car~  1.94e+3       NA       NA        NA        NA 
#>  9 Good  ran_pa~ color  cor__(I~ -9.58e-1       NA       NA        NA        NA 
#> 10 Good  ran_pa~ Resid~ sd__Obs~  1.36e+3       NA       NA        NA        NA 
#> # ... with 15 more rows