Retrieve tidy dataframe if it exists.

Check if a tidy method exits for a given object, either in broom or in broom.mixed. If it does, return the model summary dataframe, if not, return a NULL.

tidy(x, ...)

Arguments

x
An object to be converted into a tidy tibble::tibble().

...
Additional arguments to tidying method.

Examples

library(lme4)

# mixed-effects models (`broom.mixed` will be used)
lmm.mod <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
broomExtra::tidy(lmm.mod, effects = "fixed")

#> # A tibble: 2 x 5
#>   effect term        estimate std.error statistic
#>   <chr>  <chr>          <dbl>     <dbl>     <dbl>
#> 1 fixed  (Intercept)    251.       6.82     36.8 
#> 2 fixed  Days            10.5      1.55      6.77

# linear model (`broom` will be used)
lm.mod <- lm(Reaction ~ Days, sleepstudy)
broomExtra::tidy(lm.mod, exponentiate = TRUE, conf.int = TRUE)

#> Warning: Exponentiating coefficients, but model did not use a log or logit link function.

#> # A tibble: 2 x 7
#>   term        estimate std.error statistic  p.value conf.low conf.high
#>   <chr>          <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl>
#> 1 (Intercept) 1.53e109      6.61     38.0  2.16e-87 3.30e103  7.06e114
#> 2 Days        3.51e  4      1.24      8.45 9.89e-15 3.05e  3  4.05e  5

# anova (`broom` will be used)
broomExtra::tidy(stats::anova(stats::lm(wt ~ am, mtcars)))

#> # A tibble: 2 x 6
#>   term         df sumsq meansq statistic    p.value
#>   <chr>     <int> <dbl>  <dbl>     <dbl>      <dbl>
#> 1 am            1  14.2 14.2        27.6  0.0000113
#> 2 Residuals    30  15.4  0.515      NA   NA