Retrieve augmented dataframe if it exists.

Check if a augment method exits for a given object, either in broom or in broom.mixed. If it does, return the model summary dataframe, if not, return a NULL.

augment(x, ...)

Arguments

x
Model object or other R object with information to append to observations.

...
Addition arguments to augment method.

Value

A tibble::tibble() with information about data points.

Methods

See the following help topics for more details about individual methods:broom

augment.betamfx: betamfx

augment.betareg: betareg

augment.clm: clm

augment.coxph: coxph

augment.decomposed.ts: decomposed.ts

augment.drc: drc

augment.factanal: factanal

augment.felm: felm

augment.fixest: fixest

augment.glm: glm

augment.glmRob: glmRob

augment.htest: htest

augment.ivreg: ivreg

augment.kmeans: kmeans

augment.lm: lm

augment.lmRob: lmRob

augment.loess: loess

augment.Mclust: Mclust

augment.mfx: mfx

augment.mjoint: mjoint

augment.nlrq: nlrq

augment.nls: nls

augment.pam: pam

augment.plm: plm

augment.poLCA: poLCA

augment.polr: polr

augment.prcomp: prcomp

augment.rlm: rlm

augment.rma: rma

augment.robustbase.glmrob: glmrob

augment.robustbase.lmrob: lmrob

augment.rq: rq

augment.rqs: rqs

augment.sarlm: sarlm

augment.smooth.spline: smooth.spline

augment.speedlm: speedlm

augment.stl: stl

augment.survreg: survreg

null_tidiers: NULL

broom.mixed

augment.ranef.mer: ranef.mer

brms_tidiers: brmsfit

glmmadmb_tidiers: glmmadmb

glmmTMB_tidiers: glmmTMB

lme4_tidiers: merMod

nlme_tidiers: gls, lme

See also

Examples

set.seed(123)
library(lme4)

#> Loading required package: Matrix

# mixed-effects models (`broom.mixed` will be used)
lmm.mod <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
broomExtra::augment(lmm.mod)

#> # A tibble: 180 x 14
#>    Reaction  Days Subject .fitted  .resid   .hat .cooksd .fixed   .mu .offset
#>       <dbl> <dbl> <fct>     <dbl>   <dbl>  <dbl>   <dbl>  <dbl> <dbl>   <dbl>
#>  1     250.     0 308        254.   -4.10 0.229  0.00496   251.  254.       0
#>  2     259.     1 308        273.  -14.6  0.170  0.0402    262.  273.       0
#>  3     251.     2 308        293.  -42.2  0.127  0.226     272.  293.       0
#>  4     321.     3 308        313.    8.78 0.101  0.00731   283.  313.       0
#>  5     357.     4 308        332.   24.5  0.0910 0.0506    293.  332.       0
#>  6     415.     5 308        352.   62.7  0.0981 0.362     304.  352.       0
#>  7     382.     6 308        372.   10.5  0.122  0.0134    314.  372.       0
#>  8     290.     7 308        391. -101.   0.162  1.81      325.  391.       0
#>  9     431.     8 308        411.   19.6  0.219  0.106     335.  411.       0
#> 10     466.     9 308        431.   35.7  0.293  0.571     346.  431.       0
#>    .sqrtXwt .sqrtrwt .weights  .wtres
#>       <dbl>    <dbl>    <dbl>   <dbl>
#>  1        1        1        1   -4.10
#>  2        1        1        1  -14.6 
#>  3        1        1        1  -42.2 
#>  4        1        1        1    8.78
#>  5        1        1        1   24.5 
#>  6        1        1        1   62.7 
#>  7        1        1        1   10.5 
#>  8        1        1        1 -101.  
#>  9        1        1        1   19.6 
#> 10        1        1        1   35.7 
#> # ... with 170 more rows

# linear model (`broom` will be used)
lm.mod <- lm(Reaction ~ Days, sleepstudy)
broomExtra::augment(lm.mod)

#> # A tibble: 180 x 8
#>    Reaction  Days .fitted .resid .std.resid    .hat .sigma   .cooksd
#>       <dbl> <dbl>   <dbl>  <dbl>      <dbl>   <dbl>  <dbl>     <dbl>
#>  1     250.     0    251.  -1.85    -0.0390 0.0192    47.8 0.0000149
#>  2     259.     1    262.  -3.17    -0.0669 0.0138    47.8 0.0000313
#>  3     251.     2    272. -21.5     -0.454  0.00976   47.8 0.00101  
#>  4     321.     3    283.  38.6      0.813  0.00707   47.8 0.00235  
#>  5     357.     4    293.  63.6      1.34   0.00572   47.6 0.00514  
#>  6     415.     5    304. 111.       2.33   0.00572   47.1 0.0157   
#>  7     382.     6    314.  68.0      1.43   0.00707   47.6 0.00728  
#>  8     290.     7    325. -34.5     -0.727  0.00976   47.8 0.00261  
#>  9     431.     8    335.  95.4      2.01   0.0138    47.3 0.0284   
#> 10     466.     9    346. 121.       2.56   0.0192    47.0 0.0639   
#> # ... with 170 more rows