Bootstrapped dataframe with augmented predictions from each sample.

Bootstrapped dataframe with augmented predictions from each sample.

boot_augment(data, times = 25, strata = NULL, apparent = FALSE, ..f,
  ..., augment.args = list())

Arguments

data
Dataframe (or tibble) from which variables are to be taken.

times
The number of bootstrap samples.

strata
A variable that is used to conduct stratified sampling. When not NULL, each bootstrap sample is created within the stratification variable.

apparent
A logical. Should an extra resample be added where the analysis and holdout subset are the entire data set. This is required for some estimators used by the summary function that require the apparent error rate.

..f
A function, or function name as a string.

...
Arguments to function.

These dots support tidy-dots features.

augment.args
A list of arguments to be used in the relevant S3 method.

Value

A tibble::tibble() with information about data points.

Methods

See the following help topics for more details about individual methods: broom

augment.betareg: betareg

augment.clm: clm

augment.coxph: coxph

augment.decomposed.ts: decomposed.ts

augment.drc: drc

augment.factanal: factanal

augment.felm: felm

augment.glm: glm

augment.glmRob: glmRob

augment.htest: htest

augment.ivreg: ivreg

augment.kmeans: kmeans

augment.lm: lm

augment.lmRob: lmRob

augment.loess: loess

augment.Mclust: Mclust

augment.mjoint: mjoint

augment.nlrq: nlrq

augment.nls: nls

augment.pam: pam

augment.plm: plm

augment.poLCA: poLCA

augment.polr: polr

augment.prcomp: prcomp

augment.rlm: rlm

augment.rma: rma

augment.rq: rq

augment.rqs: rqs

augment.smooth.spline: smooth.spline

augment.speedlm: speedlm

augment.stl: stl

augment.survreg: survreg

null_tidiers: NULL

broom.mixed

augment.ranef.mer: ranef.mer

brms_tidiers: brmsfit

glmmadmb_tidiers: glmmadmb

glmmTMB_tidiers: glmmTMB

lme4_tidiers: merMod

nlme_tidiers: gls, lme

Examples

set.seed(123)

# example-1: linear model
broomExtra::boot_augment(
  data = mtcars,
  times = 10,
  ..f = stats::lm,
  formula = mpg ~ wt,
  na.action = na.omit
)

#> # A tibble: 320 x 10
#>    id     .rownames   mpg    wt .fitted  .resid .std.resid   .hat .sigma .cooksd
#>    <chr>  <chr>     <dbl> <dbl>   <dbl>   <dbl>      <dbl>  <dbl>  <dbl>   <dbl>
#>  1 Boots~ Maserati~  15    3.57   18.0   2.96      -1.12   0.0458   2.69 3.01e-2
#>  2 Boots~ Cadillac~  10.4  5.25    7.71 -2.69       1.15   0.250    2.69 2.18e-1
#>  3 Boots~ Honda Ci~  30.4  1.62   29.9  -0.522      0.205  0.111    2.75 2.61e-3
#>  4 Boots~ Merc 450~  15.2  3.78   16.7   1.48      -0.563  0.0582   2.74 9.78e-3
#>  5 Boots~ Datsun 7~  22.8  2.32   25.6   2.78      -1.05   0.0500   2.70 2.93e-2
#>  6 Boots~ Merc 280   19.2  3.44   18.8  -0.449      0.170  0.0401   2.75 6.00e-4
#>  7 Boots~ Fiat 128   32.4  2.2    26.3  -6.09       2.32   0.0574   2.49 1.64e-1
#>  8 Boots~ Dodge Ch~  15.5  3.52   18.3   2.76      -1.04   0.0435   2.70 2.48e-2
#>  9 Boots~ Merc 280C  17.8  3.44   18.8   0.951     -0.359  0.0401   2.75 2.68e-3
#> 10 Boots~ Hornet S~  18.7  3.44   18.8   0.0506    -0.0191 0.0401   2.75 7.61e-6
#> # ... with 310 more rows

# example-2: linear mixed-effects model
library(lme4)

broomExtra::boot_augment(
  data = sleepstudy,
  times = 25,
  ..f = lme4::lmer,
  formula = Reaction ~ Days + (Days | Subject)
)

#> Warning: Model failed to converge with max|grad| = 0.00416738 (tol = 0.002, component 1)

#> Warning: Model failed to converge with max|grad| = 0.344884 (tol = 0.002, component 1)

#> Warning: Model failed to converge with max|grad| = 0.00307876 (tol = 0.002, component 1)

#> # A tibble: 4,500 x 16
#>    id    .rownames Reaction  Days Subject .fitted .resid   .hat .cooksd .fixed
#>    <chr> <chr>        <dbl> <dbl> <fct>     <dbl>  <dbl>  <dbl>   <dbl>  <dbl>
#>  1 Boot~ 145           315.     4 369        305.   9.37 0.0689 0.00686   297.
#>  2 Boot~ 165           279.     4 371        293. -13.5  0.0689 0.0143    297.
#>  3 Boot~ 57            454.     6 332        341. 114.   0.164  2.97      319.
#>  4 Boot~ 148           348.     7 369        336.  12.2  0.157  0.0321    330.
#>  5 Boot~ 163           278.     2 371        269.   8.61 0.124  0.0117    275.
#>  6 Boot~ 161           270.     0 371        246.  24.0  0.223  0.208     253.
#>  7 Boot~ 66            338.     5 333        324.  14.0  0.109  0.0265    308.
#>  8 Boot~ 4             321.     3 308        309.  12.4  0.155  0.0331    286.
#>  9 Boot~ 74            255.     3 334        262.  -7.52 0.134  0.00994   286.
#> 10 Boot~ 133           327.     2 352        323.   3.46 0.213  0.00404   275.
#> # ... with 4,490 more rows, and 6 more variables: .mu <dbl>, .offset <dbl>,
#> #   .sqrtXwt <dbl>, .sqrtrwt <dbl>, .weights <dbl>, .wtres <dbl>