Retrieve model summary dataframe if it exists.

Check if a glance method exits for a given object, either in broom or in broom.mixed. If it does, return the model summary dataframe, if not, return a NULL. In this case, you can try the glance_performance() function.

glance(x, ...)

Arguments

x
model or other R object to convert to single-row data frame

...
other arguments passed to methods

Methods

See the following help topics for more details about individual methods:broom

data.frame_tidiers: data.frame

durbinWatsonTest_tidiers: durbinWatsonTest

glance.aareg: aareg

glance.aov: aov

glance.Arima: Arima

glance.betamfx: betamfx

glance.betareg: betareg

glance.biglm: biglm

glance.binDesign: binDesign

glance.cch: cch

glance.clm: clm

glance.clmm: clmm

glance.coeftest: coeftest

glance.coxph: coxph

glance.cv.glmnet: cv.glmnet

glance.drc: drc

glance.ergm: ergm

glance.factanal: factanal

glance.felm: felm

glance.fitdistr: fitdistr

glance.fixest: fixest

glance.gam: gam

glance.garch: garch

glance.geeglm: geeglm

glance.glm: glm

glance.glmnet: glmnet

glance.glmRob: glmRob

glance.gmm: gmm

glance.ivreg: ivreg

glance.kmeans: kmeans

glance.lavaan: lavaan

glance.lm: lm

glance.lmodel2: lmodel2

glance.lmRob: lmRob

glance.margins: margins

glance.Mclust: Mclust

glance.mfx: logitmfx, mfx, negbinmfx, poissonmfx, probitmfx

glance.mjoint: mjoint

glance.mlogit: mlogit

glance.muhaz: muhaz

glance.multinom: multinom

glance.nlrq: nlrq

glance.nls: nls

glance.orcutt: orcutt

glance.pam: pam

glance.plm: plm

glance.poLCA: poLCA

glance.polr: polr

glance.pyears: pyears

glance.ridgelm: ridgelm

glance.rlm: rlm

glance.rma: rma

glance.robustbase.lmrob: lmrob

glance.rq: rq

glance.sarlm: sarlm

glance.smooth.spline: smooth.spline

glance.speedglm: speedglm

glance.speedlm: speedlm

glance.survdiff: survdiff

glance.survexp: survexp

glance.survfit: survfit

glance.survreg: survreg

glance.svyglm: svyglm

glance.svyolr: svyolr

list_tidiers: list

null_tidiers: NULL

summary_tidiers: summaryDefault

tidy.htest: htest

broom.mixed

brms_tidiers: brmsfit

glmmadmb_tidiers: glmmadmb

glmmTMB_tidiers: glmmTMB

lme4_tidiers: merMod

nlme_tidiers: lme

rstanarm_tidiers: stanreg

See also

Examples

set.seed(123)
lm.mod <- lm(Sepal.Length ~ Species, iris)
broomExtra::glance(lm.mod)

#> # A tibble: 1 x 12
#>   r.squared adj.r.squared sigma statistic  p.value    df logLik   AIC   BIC
#>       <dbl>         <dbl> <dbl>     <dbl>    <dbl> <dbl>  <dbl> <dbl> <dbl>
#> 1     0.619         0.614 0.515      119. 1.67e-31     2  -112.  231.  243.
#>   deviance df.residual  nobs
#>      <dbl>       <int> <int>
#> 1     39.0         147   150