Augmented data from grouped analysis of any function that has data argument in its function call.

Augmented data from grouped analysis of any function that has data argument in its function call.

grouped_augment(data, grouping.vars, ..f, ..., augment.args = list())

Arguments

data
Dataframe (or tibble) from which variables are to be taken.

grouping.vars
Grouping variables.

..f
A function, or function name as a string.

...
Arguments to function.

These dots support tidy-dots features.

augment.args
A list of arguments to be used in the relevant S3 method.

Value

A tibble::tibble() with information about data points.

Methods

See the following help topics for more details about individual methods: broom

augment.betareg: betareg

augment.coxph: coxph

augment.decomposed.ts: decomposed.ts

augment.factanal: factanal

augment.felm: felm

augment.glm: glm

augment.glmRob: glmRob

augment.htest: htest

augment.ivreg: ivreg

augment.kmeans: kmeans

augment.lm: lm

augment.lmRob: lmRob

augment.loess: loess

augment.Mclust: Mclust

augment.mjoint: mjoint

augment.nlrq: nlrq

augment.nls: nls

augment.pam: pam

augment.plm: plm

augment.poLCA: poLCA

augment.prcomp: prcomp

augment.rlm: rlm

augment.rq: rq

augment.rqs: rqs

augment.smooth.spline: smooth.spline

augment.speedlm: speedlm

augment.stl: stl

augment.survreg: survreg

null_tidiers: NULL

ordinal_tidiers: clm, polr

broom.mixed

augment.ranef.mer: ranef.mer

brms_tidiers: brmsfit

glmmadmb_tidiers: glmmadmb

glmmTMB_tidiers: glmmTMB

lme4_tidiers: merMod

nlme_tidiers: lme

Examples

set.seed(123)
# to speed up computation, let's use only 50% of the data

# linear model
broomExtra::grouped_augment(
  data = dplyr::sample_frac(tbl = ggplot2::diamonds, size = 0.5),
  grouping.vars = c(cut, color),
  formula = price ~ carat - 1,
  ..f = stats::lm,
  na.action = na.omit
)

#> # A tibble: 26,970 x 10
#>    cut   color price carat .fitted .resid .std.resid    .hat .sigma  .cooksd
#>    <ord> <ord> <int> <dbl>   <dbl>  <dbl>      <dbl>   <dbl>  <dbl>    <dbl>
#>  1 Fair  D      1323  0.5    2623.  1300.     -0.701 0.00311  1862. 0.00154 
#>  2 Fair  D      3382  0.9    4722.  1340.     -0.725 0.0101   1862. 0.00536 
#>  3 Fair  D       675  0.42   2203.  1528.     -0.824 0.00220  1860. 0.00150 
#>  4 Fair  D      8716  1.01   5299. -3417.      1.85  0.0127   1829. 0.0442  
#>  5 Fair  D      1806  0.7    3672.  1866.     -1.01  0.00610  1856. 0.00624 
#>  6 Fair  D      3304  1      5246.  1942.     -1.05  0.0125   1855. 0.0140  
#>  7 Fair  D      4751  1.01   5299.   548.     -0.297 0.0127   1867. 0.00113 
#>  8 Fair  D      3296  0.7    3672.   376.     -0.203 0.00610  1867. 0.000254
#>  9 Fair  D      2518  0.76   3987.  1469.     -0.794 0.00719  1861. 0.00457 
#> 10 Fair  D     10562  1      5246. -5316.      2.88  0.0125   1773. 0.105   
#> # ... with 26,960 more rows

# linear mixed effects model
broomExtra::grouped_augment(
  data = dplyr::sample_frac(tbl = ggplot2::diamonds, size = 0.5),
  grouping.vars = cut,
  ..f = lme4::lmer,
  formula = price ~ carat + (carat | color) - 1,
  control = lme4::lmerControl(optimizer = "bobyqa")
)

#> # A tibble: 26,970 x 15
#>    cut   price carat color .fitted .resid    .hat .cooksd .fixed    .mu .offset
#>    <ord> <int> <dbl> <ord>   <dbl>  <dbl>   <dbl>   <dbl>  <dbl>  <dbl>   <dbl>
#>  1 Fair   2848  0.75 D      3144.   -296. 0.0117  3.13e-4  2850. 3.14e3       0
#>  2 Fair   5268  1.5  J      5765.   -497. 0.0145  1.10e-3  5699. 5.76e3       0
#>  3 Fair   4205  1.01 D      4994.   -789. 0.0108  2.06e-3  3838. 4.99e3       0
#>  4 Fair   1633  0.68 G      2226.   -593. 0.00915 9.78e-4  2584. 2.23e3       0
#>  5 Fair   3975  1    F      4426.   -451. 0.00640 3.94e-4  3800. 4.43e3       0
#>  6 Fair   3876  1.03 E      4665.   -789. 0.0110  2.10e-3  3914. 4.67e3       0
#>  7 Fair   4072  1.01 G      4229.   -157. 0.00576 4.30e-5  3838. 4.23e3       0
#>  8 Fair    527  0.37 F        83.3   444. 0.0172  1.05e-3  1406. 8.33e1       0
#>  9 Fair  15030  2.48 I     12504.   2526. 0.104   2.47e-1  9423. 1.25e4       0
#> 10 Fair   4791  1.01 F      4495.    296. 0.00646 1.71e-4  3838. 4.50e3       0
#> # ... with 26,960 more rows, and 4 more variables: .sqrtXwt <dbl>,
#> #   .sqrtrwt <dbl>, .weights <dbl>, .wtres <dbl>