Augmented data from grouped analysis of any function that has data argument in its function call.

Augmented data from grouped analysis of any function that has data argument in its function call.

grouped_augment(data, grouping.vars, ..f, ..., augment.args = list())

Arguments

data
Dataframe (or tibble) from which variables are to be taken.

grouping.vars
Grouping variables.

..f
A function, or function name as a string.

...
<dynamic> Arguments for .fn.

augment.args
A list of arguments to be used in the relevant S3 method.

Value

A tibble::tibble() with information about data points.

Methods

See the following help topics for more details about individual methods:broom

augment.betareg: betareg

augment.clm: clm

augment.coxph: coxph

augment.decomposed.ts: decomposed.ts

augment.drc: drc

augment.factanal: factanal

augment.felm: felm

augment.glm: glm

augment.glmRob: glmRob

augment.htest: htest

augment.ivreg: ivreg

augment.kmeans: kmeans

augment.lm: lm

augment.lmRob: lmRob

augment.loess: loess

augment.Mclust: Mclust

augment.mjoint: mjoint

augment.nlrq: nlrq

augment.nls: nls

augment.pam: pam

augment.plm: plm

augment.poLCA: poLCA

augment.polr: polr

augment.prcomp: prcomp

augment.rlm: rlm

augment.rma: rma

augment.robustbase.glmrob: glmrob

augment.robustbase.lmrob: lmrob

augment.rq: rq

augment.rqs: rqs

augment.smooth.spline: smooth.spline

augment.speedlm: speedlm

augment.stl: stl

augment.survreg: survreg

null_tidiers: NULL

broom.mixed

augment.ranef.mer: ranef.mer

brms_tidiers: brmsfit

glmmadmb_tidiers: glmmadmb

glmmTMB_tidiers: glmmTMB

lme4_tidiers: merMod

nlme_tidiers: gls, lme

See also

Examples

set.seed(123)
# to speed up computation, let's use only 50% of the data

# linear model
broomExtra::grouped_augment(
  data = dplyr::sample_frac(tbl = ggplot2::diamonds, size = 0.5),
  grouping.vars = c(cut, color),
  formula = price ~ carat - 1,
  ..f = stats::lm,
  na.action = na.omit,
  augment.args = list(se_fit = TRUE)
)

#> # A tibble: 26,970 x 11
#>    cut   color price carat .fitted .se.fit .resid .std.resid    .hat .sigma
#>    <ord> <ord> <int> <dbl>   <dbl>   <dbl>  <dbl>      <dbl>   <dbl>  <dbl>
#>  1 Fair  D      1323  0.5    2623.   104.   1300.     -0.701 0.00311  1862.
#>  2 Fair  D      3382  0.9    4722.   186.   1340.     -0.725 0.0101   1862.
#>  3 Fair  D       675  0.42   2203.    87.0  1528.     -0.824 0.00220  1860.
#>  4 Fair  D      8716  1.01   5299.   209.  -3417.      1.85  0.0127   1829.
#>  5 Fair  D      1806  0.7    3672.   145.   1866.     -1.01  0.00610  1856.
#>  6 Fair  D      3304  1      5246.   207.   1942.     -1.05  0.0125   1855.
#>  7 Fair  D      4751  1.01   5299.   209.    548.     -0.297 0.0127   1867.
#>  8 Fair  D      3296  0.7    3672.   145.    376.     -0.203 0.00610  1867.
#>  9 Fair  D      2518  0.76   3987.   157.   1469.     -0.794 0.00719  1861.
#> 10 Fair  D     10562  1      5246.   207.  -5316.      2.88  0.0125   1773.
#>     .cooksd
#>       <dbl>
#>  1 0.00154 
#>  2 0.00536 
#>  3 0.00150 
#>  4 0.0442  
#>  5 0.00624 
#>  6 0.0140  
#>  7 0.00113 
#>  8 0.000254
#>  9 0.00457 
#> 10 0.105   
#> # ... with 26,960 more rows

# linear mixed effects model
broomExtra::grouped_augment(
  data = dplyr::sample_frac(tbl = ggplot2::diamonds, size = 0.5),
  grouping.vars = "cut",
  ..f = lme4::lmer,
  formula = price ~ carat + (carat | color) - 1,
  control = lme4::lmerControl(optimizer = "bobyqa")
)

#> # A tibble: 26,970 x 15
#>    cut   price carat color .fitted .resid    .hat   .cooksd .fixed     .mu
#>    <ord> <int> <dbl> <ord>   <dbl>  <dbl>   <dbl>     <dbl>  <dbl>   <dbl>
#>  1 Fair   2848  0.75 D      3144.   -296. 0.0117  0.000313   2850.  3144. 
#>  2 Fair   5268  1.5  J      5765.   -497. 0.0145  0.00110    5699.  5765. 
#>  3 Fair   4205  1.01 D      4994.   -789. 0.0108  0.00206    3838.  4994. 
#>  4 Fair   1633  0.68 G      2226.   -593. 0.00915 0.000978   2584.  2226. 
#>  5 Fair   3975  1    F      4426.   -451. 0.00640 0.000394   3800.  4426. 
#>  6 Fair   3876  1.03 E      4665.   -789. 0.0110  0.00210    3914.  4665. 
#>  7 Fair   4072  1.01 G      4229.   -157. 0.00576 0.0000430  3838.  4229. 
#>  8 Fair    527  0.37 F        83.3   444. 0.0172  0.00105    1406.    83.3
#>  9 Fair  15030  2.48 I     12504.   2526. 0.104   0.247      9423. 12504. 
#> 10 Fair   4791  1.01 F      4495.    296. 0.00646 0.000171   3838.  4495. 
#>    .offset .sqrtXwt .sqrtrwt .weights .wtres
#>      <dbl>    <dbl>    <dbl>    <dbl>  <dbl>
#>  1       0        1        1        1  -296.
#>  2       0        1        1        1  -497.
#>  3       0        1        1        1  -789.
#>  4       0        1        1        1  -593.
#>  5       0        1        1        1  -451.
#>  6       0        1        1        1  -789.
#>  7       0        1        1        1  -157.
#>  8       0        1        1        1   444.
#>  9       0        1        1        1  2526.
#> 10       0        1        1        1   296.
#> # ... with 26,960 more rows