Augmented data from grouped analysis of any function that has data argument in its function call.

Augmented data from grouped analysis of any function that has data argument in its function call.

grouped_augment(data, grouping.vars, ..f, ..., augment.args = list())

Arguments

data
Dataframe (or tibble) from which variables are to be taken.

grouping.vars
Grouping variables.

..f
A function, or function name as a string.

...
<dynamic> Arguments for .fn.

augment.args
A list of arguments to be used in the relevant S3 method.

Value

A tibble::tibble() with information about data points.

Methods

See the following help topics for more details about individual methods:broom

augment.betamfx: betamfx

augment.betareg: betareg

augment.clm: clm

augment.coxph: coxph

augment.decomposed.ts: decomposed.ts

augment.drc: drc

augment.factanal: factanal

augment.felm: felm

augment.fixest: fixest

augment.glm: glm

augment.glmRob: glmRob

augment.htest: htest

augment.ivreg: ivreg

augment.kmeans: kmeans

augment.lm: lm

augment.lmRob: lmRob

augment.loess: loess

augment.Mclust: Mclust

augment.mfx: logitmfx, mfx, negbinmfx, poissonmfx, probitmfx

augment.mjoint: mjoint

augment.mlogit: mlogit

augment.nlrq: nlrq

augment.nls: nls

augment.pam: pam

augment.plm: plm

augment.poLCA: poLCA

augment.polr: polr

augment.prcomp: prcomp

augment.rlm: rlm

augment.rma: rma

augment.robustbase.glmrob: glmrob

augment.robustbase.lmrob: lmrob

augment.rq: rq

augment.rqs: rqs

augment.sarlm: sarlm

augment.smooth.spline: smooth.spline

augment.speedlm: speedlm

augment.stl: stl

augment.survreg: survreg

data.frame_tidiers: data.frame

null_tidiers: NULL

broom.mixed

augment.ranef.mer: ranef.mer

brms_tidiers: brmsfit

glmmadmb_tidiers: glmmadmb

glmmTMB_tidiers: glmmTMB

lme4_tidiers: merMod

nlme_tidiers: gls, lme

See also

Examples

set.seed(123)

# linear mixed effects model
broomExtra::grouped_augment(
  data = dplyr::mutate(MASS::Aids2, interval = death - diag),
  grouping.vars = sex,
  ..f = lme4::lmer,
  formula = interval ~ age + (1 | status),
  control = lme4::lmerControl(optimizer = "bobyqa")
)

#> # A tibble: 2,843 x 15
#>    sex   interval   age status .fitted  .resid   .hat    .cooksd .fixed   .mu
#>    <fct>    <int> <int> <fct>    <dbl>   <dbl>  <dbl>      <dbl>  <dbl> <dbl>
#>  1 F          543    30 A         604.  -61.5  0.0261 0.000301     493.  604.
#>  2 F          138    44 D         321. -183.   0.0181 0.00182      432.  321.
#>  3 F            1    25 D         403. -402.   0.0289 0.0144       515.  403.
#>  4 F           43    55 D         273. -230.   0.0249 0.00402      385.  273.
#>  5 F          215    70 D         208.    6.90 0.0495 0.00000756   319.  208.
#>  6 F          260    17 D         438. -178.   0.0420 0.00421      549.  438.
#>  7 F          511    57 D         265.  246.   0.0271 0.00505      376.  265.
#>  8 F           92    71 D         204. -112.   0.0517 0.00208      315.  204.
#>  9 F          494    47 D         308.  186.   0.0190 0.00198      419.  308.
#> 10 F           49    67 D         221. -172.   0.0431 0.00404      333.  221.
#>    .offset .sqrtXwt .sqrtrwt .weights  .wtres
#>      <dbl>    <dbl>    <dbl>    <dbl>   <dbl>
#>  1       0        1        1        1  -61.5 
#>  2       0        1        1        1 -183.  
#>  3       0        1        1        1 -402.  
#>  4       0        1        1        1 -230.  
#>  5       0        1        1        1    6.90
#>  6       0        1        1        1 -178.  
#>  7       0        1        1        1  246.  
#>  8       0        1        1        1 -112.  
#>  9       0        1        1        1  186.  
#> 10       0        1        1        1 -172.  
#> # ... with 2,833 more rows