Model summary output from grouped analysis of any function that has data argument in its function call.

Model summary output from grouped analysis of any function that has data argument in its function call.

grouped_glance(data, grouping.vars, ..f, ...)

Arguments

data
Dataframe (or tibble) from which variables are to be taken.

grouping.vars
Grouping variables.

..f
A function, or function name as a string.

...
Arguments to function.

These dots support tidy-dots features.

Methods

See the following help topics for more details about individual methods: broom

durbinWatsonTest_tidiers: durbinWatsonTest

glance.aareg: aareg

glance.aov: aov

glance.Arima: Arima

glance.betareg: betareg

glance.biglm: biglm

glance.binDesign: binDesign

glance.cch: cch

glance.clm: clm

glance.clmm: clmm

glance.coxph: coxph

glance.cv.glmnet: cv.glmnet

glance.ergm: ergm

glance.factanal: factanal

glance.felm: felm

glance.fitdistr: fitdistr

glance.Gam: Gam

glance.garch: garch

glance.glm: glm

glance.glmnet: glmnet

glance.glmRob: glmRob

glance.gmm: gmm

glance.ivreg: ivreg

glance.kmeans: kmeans

glance.lavaan: lavaan

glance.lm: lm

glance.lmodel2: lmodel2

glance.lmRob: lmRob

glance.Mclust: Mclust

glance.mjoint: mjoint

glance.muhaz: muhaz

glance.multinom: multinom

glance.nlrq: nlrq

glance.nls: nls

glance.orcutt: orcutt

glance.pam: pam

glance.plm: plm

glance.poLCA: poLCA

glance.polr: polr

glance.pyears: pyears

glance.ridgelm: ridgelm

glance.rlm: rlm

glance.rq: rq

glance.smooth.spline: smooth.spline

glance.speedlm: speedlm

glance.survdiff: survdiff

glance.survexp: survexp

glance.survfit: survfit

glance.survreg: survreg

glance.svyolr: svyolr

list_tidiers: list

mgcv_glance_gam: gam

null_tidiers: NULL

tidy.htest: htest

broom.mixed

brms_tidiers: brmsfit

glmmadmb_tidiers: glmmadmb

glmmTMB_tidiers: glmmTMB

lme4_tidiers: merMod

nlme_tidiers: lme

rstanarm_tidiers: stanreg

Examples

set.seed(123)
# to speed up computation, let's use only 50% of the data

# linear model
broomExtra::grouped_glance(
  data = dplyr::sample_frac(tbl = ggplot2::diamonds, size = 0.5),
  grouping.vars = c(cut, color),
  formula = price ~ carat - 1,
  ..f = stats::lm,
  na.action = na.omit
)

#> # A tibble: 35 x 14
#>    cut   color r.squared adj.r.squared sigma statistic   p.value    df logLik
#>    <ord> <ord>     <dbl>         <dbl> <dbl>     <dbl>     <dbl> <dbl>  <dbl>
#>  1 Fair  D         0.884         0.883 1857.      641. 4.45e- 41     1  -760.
#>  2 Fair  E         0.876         0.875 1370.      708. 3.52e- 47     1  -872.
#>  3 Fair  F         0.874         0.873 1989.     1071. 1.68e- 71     1 -1406.
#>  4 Fair  G         0.849         0.848 2138.      887. 1.03e- 66     1 -1444.
#>  5 Fair  H         0.876         0.875 2412.      998. 7.68e- 66     1 -1307.
#>  6 Fair  I         0.915         0.914 1499.      850. 4.86e- 44     1  -698.
#>  7 Fair  J         0.885         0.883 2189.      416. 4.80e- 27     1  -501.
#>  8 Good  D         0.860         0.860 1729.     2065. 2.66e-145     1 -2981.
#>  9 Good  E         0.870         0.870 1674.     3084. 2.50e-206     1 -4084.
#> 10 Good  F         0.873         0.873 1677.     3110. 1.76e-204     1 -3997.
#> # ... with 25 more rows, and 5 more variables: AIC <dbl>, BIC <dbl>,
#> #   deviance <dbl>, df.residual <int>, nobs <int>

# linear mixed effects model
broomExtra::grouped_glance(
  data = dplyr::sample_frac(tbl = ggplot2::diamonds, size = 0.5),
  grouping.vars = cut,
  ..f = lme4::lmer,
  formula = price ~ carat + (carat | color) - 1,
  control = lme4::lmerControl(optimizer = "bobyqa")
)

#> # A tibble: 5 x 7
#>   cut       sigma  logLik     AIC     BIC REMLcrit df.residual
#>   <ord>     <dbl>   <dbl>   <dbl>   <dbl>    <dbl>       <int>
#> 1 Fair      1830.  -7257.  14525.  14548.   14515.         806
#> 2 Good      1373. -21027.  42064.  42093.   42054.        2425
#> 3 Very Good 1362. -51577. 103165. 103198.  103155.        5964
#> 4 Premium   1557. -60736. 121482. 121516.  121472.        6917
#> 5 Ideal     1257. -92766. 185542. 185579.  185532.       10833