Model summary output from grouped analysis of any function that has data argument in its function call.

Model summary output from grouped analysis of any function that has data argument in its function call.

grouped_glance(data, grouping.vars, ..f, ...)

Arguments

data
Dataframe (or tibble) from which variables are to be taken.

grouping.vars
Grouping variables.

..f
A function, or function name as a string.

...
<dynamic> Arguments for .fn.

See also

Examples

set.seed(123)

# linear mixed effects model
broomExtra::grouped_glance(
  data = dplyr::mutate(MASS::Aids2, interval = death - diag),
  grouping.vars = sex,
  ..f = lme4::lmer,
  formula = interval ~ age + (1 | status),
  control = lme4::lmerControl(optimizer = "bobyqa")
)

#> # A tibble: 2 x 7
#>   sex   sigma  logLik    AIC    BIC REMLcrit df.residual
#>   <fct> <dbl>   <dbl>  <dbl>  <dbl>    <dbl>       <int>
#> 1 F      415.   -656.  1321.  1331.    1313.          85
#> 2 M      355. -20079. 40165. 40189.   40157.        2750