Model summary output from grouped analysis of any function that has data argument in its function call.

Model summary output from grouped analysis of any function that has data argument in its function call.

grouped_glance(data, grouping.vars, ..f, ...)

Arguments

data
Dataframe (or tibble) from which variables are to be taken.

grouping.vars
Grouping variables.

..f
A function, or function name as a string.

...
<dynamic> Arguments for .fn.

Note

For available methods, see- https://indrajeetpatil.github.io/broomExtra/articles/available_methods.html

Methods

See the following help topics for more details about individual methods:broom

durbinWatsonTest_tidiers: durbinWatsonTest

glance.aareg: aareg

glance.aov: aov

glance.Arima: Arima

glance.betareg: betareg

glance.biglm: biglm

glance.binDesign: binDesign

glance.cch: cch

glance.clm: clm

glance.clmm: clmm

glance.coxph: coxph

glance.cv.glmnet: cv.glmnet

glance.drc: drc

glance.ergm: ergm

glance.factanal: factanal

glance.felm: felm

glance.fitdistr: fitdistr

glance.Gam: Gam

glance.garch: garch

glance.geeglm: geeglm

glance.glm: glm

glance.glmnet: glmnet

glance.glmRob: glmRob

glance.gmm: gmm

glance.ivreg: ivreg

glance.kmeans: kmeans

glance.lavaan: lavaan

glance.lm: lm

glance.lmodel2: lmodel2

glance.lmRob: lmRob

glance.Mclust: Mclust

glance.mjoint: mjoint

glance.muhaz: muhaz

glance.multinom: multinom

glance.nlrq: nlrq

glance.nls: nls

glance.orcutt: orcutt

glance.pam: pam

glance.plm: plm

glance.poLCA: poLCA

glance.polr: polr

glance.pyears: pyears

glance.ridgelm: ridgelm

glance.rlm: rlm

glance.rma: rma

glance.robustbase.lmrob: lmrob

glance.rq: rq

glance.smooth.spline: smooth.spline

glance.speedglm: speedglm

glance.speedlm: speedlm

glance.survdiff: survdiff

glance.survexp: survexp

glance.survfit: survfit

glance.survreg: survreg

glance.svyglm: svyglm

glance.svyolr: svyolr

list_tidiers: list

mgcv_glance_gam: gam

null_tidiers: NULL

tidy.htest: htest

broom.mixed

brms_tidiers: brmsfit

glmmadmb_tidiers: glmmadmb

glmmTMB_tidiers: glmmTMB

lme4_tidiers: merMod

nlme_tidiers: lme

rstanarm_tidiers: stanreg

See also

Examples

set.seed(123)
# to speed up computation, let's use only 50% of the data

# linear model
broomExtra::grouped_glance(
  data = dplyr::sample_frac(tbl = ggplot2::diamonds, size = 0.5),
  grouping.vars = c(cut, color),
  formula = price ~ carat - 1,
  ..f = stats::lm,
  na.action = na.omit
)

#> Error: 'grouped_glance' is not an exported object from 'namespace:broomExtra'

# linear mixed effects model
broomExtra::grouped_glance(
  data = dplyr::sample_frac(tbl = ggplot2::diamonds, size = 0.5),
  grouping.vars = "cut",
  ..f = lme4::lmer,
  formula = price ~ carat + (carat | color) - 1,
  control = lme4::lmerControl(optimizer = "bobyqa")
)

#> Error: 'grouped_glance' is not an exported object from 'namespace:broomExtra'