Model summary output from grouped analysis of any function that has data argument in its function call.

Model summary output from grouped analysis of any function that has data argument in its function call.

grouped_glance(data, grouping.vars, ..f, ...)

Arguments

data
Dataframe (or tibble) from which variables are to be taken.

grouping.vars
Grouping variables.

..f
A function, or function name as a string.

...
<dynamic> Arguments for .fn.

Methods

See the following help topics for more details about individual methods:broom

data.frame_tidiers: data.frame

durbinWatsonTest_tidiers: durbinWatsonTest

glance.aareg: aareg

glance.aov: aov

glance.Arima: Arima

glance.betamfx: betamfx

glance.betareg: betareg

glance.biglm: biglm

glance.binDesign: binDesign

glance.cch: cch

glance.clm: clm

glance.clmm: clmm

glance.coeftest: coeftest

glance.coxph: coxph

glance.crr: crr

glance.cv.glmnet: cv.glmnet

glance.drc: drc

glance.ergm: ergm

glance.factanal: factanal

glance.felm: felm

glance.fitdistr: fitdistr

glance.fixest: fixest

glance.gam: gam

glance.garch: garch

glance.geeglm: geeglm

glance.glm: glm

glance.glmnet: glmnet

glance.glmRob: glmRob

glance.gmm: gmm

glance.ivreg: ivreg

glance.kmeans: kmeans

glance.lavaan: lavaan

glance.lm: lm

glance.lmodel2: lmodel2

glance.lmRob: lmRob

glance.margins: margins

glance.Mclust: Mclust

glance.mfx: logitmfx, mfx, negbinmfx, poissonmfx, probitmfx

glance.mjoint: mjoint

glance.mlogit: mlogit

glance.muhaz: muhaz

glance.multinom: multinom

glance.negbin: negbin

glance.nlrq: nlrq

glance.nls: nls

glance.orcutt: orcutt

glance.pam: pam

glance.plm: plm

glance.poLCA: poLCA

glance.polr: polr

glance.pyears: pyears

glance.Rchoice: Rchoice

glance.ridgelm: ridgelm

glance.rlm: rlm

glance.rma: rma

glance.robustbase.lmrob: lmrob

glance.rq: rq

glance.sarlm: sarlm

glance.smooth.spline: smooth.spline

glance.speedglm: speedglm

glance.speedlm: speedlm

glance.summary.lm: summary.lm

glance.survdiff: survdiff

glance.survexp: survexp

glance.survfit: survfit

glance.survreg: survreg

glance.svyglm: svyglm

glance.svyolr: svyolr

glance.varest: varest

list_tidiers: list

null_tidiers: NULL

summary_tidiers: summaryDefault

tidy.htest: htest

broom.mixed

brms_tidiers: brmsfit

glmmadmb_tidiers: glmmadmb

glmmTMB_tidiers: glmmTMB

lme4_tidiers: merMod

nlme_tidiers: lme

rstanarm_tidiers: stanreg

See also

Examples

set.seed(123)

# linear mixed effects model
broomExtra::grouped_glance(
  data = dplyr::mutate(MASS::Aids2, interval = death - diag),
  grouping.vars = sex,
  ..f = lme4::lmer,
  formula = interval ~ age + (1 | status),
  control = lme4::lmerControl(optimizer = "bobyqa")
)

#> # A tibble: 2 x 7
#>   sex   sigma  logLik    AIC    BIC REMLcrit df.residual
#>   <fct> <dbl>   <dbl>  <dbl>  <dbl>    <dbl>       <int>
#> 1 F      415.   -656.  1321.  1331.    1313.          85
#> 2 M      355. -20079. 40165. 40189.   40157.        2750