Tidy output from grouped analysis of any function that has data argument in its function call.

Tidy output from grouped analysis of any function that has data argument in its function call.

grouped_tidy(data, grouping.vars, ..f, ..., tidy.args = list())

Arguments

data
Dataframe (or tibble) from which variables are to be taken.

grouping.vars
Grouping variables.

..f
A function, or function name as a string.

...
<dynamic> Arguments for .fn.

tidy.args
A list of arguments to be used in the relevant S3 method.

Value

A tibble::tibble() with information about model components.

Methods

See the following help topics for more details about individual methods:broom

data.frame_tidiers: data.frame

durbinWatsonTest_tidiers: durbinWatsonTest

leveneTest_tidiers: leveneTest

list_tidiers: list

metafor_tidiers: rma

null_tidiers: NULL

sp_tidiers: Line, Lines, Polygon, Polygons, SpatialLinesDataFrame, SpatialPolygons, SpatialPolygonsDataFrame

sparse_tidiers: dgCMatrix, dgTMatrix, sparseMatrix

summary_tidiers: summaryDefault

tidy.aareg: aareg

tidy.acf: acf

tidy.anova: anova

tidy.aov: aov

tidy.aovlist: aovlist

tidy.Arima: Arima

tidy.betamfx: betamfx

tidy.betareg: betareg

tidy.biglm: biglm

tidy.binDesign: binDesign

tidy.binWidth: binWidth

tidy.boot: boot

tidy.btergm: btergm

tidy.cch: cch

tidy.cld: cld

tidy.clm: clm

tidy.clmm: clmm

tidy.coeftest: coeftest

tidy.confint.glht: confint.glht

tidy.confusionMatrix: confusionMatrix

tidy.coxph: coxph

tidy.crr: crr

tidy.cv.glmnet: cv.glmnet

tidy.density: density

tidy.dist: dist

tidy.drc: drc

tidy.emmGrid: emmGrid

tidy.epi.2by2: epi.2by2

tidy.ergm: ergm

tidy.factanal: factanal

tidy.felm: felm

tidy.fitdistr: fitdistr

tidy.fixest: fixest

tidy.ftable: ftable

tidy.gam: gam

tidy.garch: garch

tidy.geeglm: geeglm

tidy.glht: glht

tidy.glm: glm

tidy.glmnet: glmnet

tidy.glmRob: glmRob

tidy.gmm: gmm

tidy.htest: htest

tidy.ivreg: ivreg

tidy.kappa: kappa

tidy.kde: kde

tidy.Kendall: Kendall

tidy.kmeans: kmeans

tidy.lavaan: lavaan

tidy.lm.beta: lm.beta

tidy.lm: lm

tidy.lmodel2: lmodel2

tidy.lmRob: lmRob

tidy.lsmobj: lsmobj

tidy.manova: manova

tidy.map: map

tidy.margins: margins

tidy.Mclust: Mclust

tidy.mediate: mediate

tidy.mfx: logitmfx, mfx, negbinmfx, poissonmfx, probitmfx

tidy.mjoint: mjoint

tidy.mle2: mle2

tidy.mlm: mlm

tidy.mlogit: mlogit

tidy.muhaz: muhaz

tidy.multinom: multinom

tidy.negbin: negbin

tidy.nlrq: nlrq

tidy.nls: nls

tidy.orcutt: orcutt

tidy.pairwise.htest: pairwise.htest

tidy.pam: pam

tidy.plm: plm

tidy.poLCA: poLCA

tidy.polr: polr

tidy.power.htest: power.htest

tidy.prcomp: prcomp

tidy.pyears: pyears

tidy.Rchoice: Rchoice

tidy.rcorr: rcorr

tidy.ref.grid: ref.grid

tidy.regsubsets: regsubsets

tidy.ridgelm: ridgelm

tidy.rlm: rlm

tidy.robustbase.glmrob: glmrob

tidy.robustbase.lmrob: lmrob

tidy.roc: roc

tidy.rq: rq

tidy.rqs: rqs

tidy.sarlm: sarlm

tidy.spec: spec

tidy.speedglm: speedglm

tidy.speedlm: speedlm

tidy.summary_emm: summary_emm

tidy.summary.glht: summary.glht

tidy.summary.lm: summary.lm

tidy.survdiff: survdiff

tidy.survexp: survexp

tidy.survfit: survfit

tidy.survreg: survreg

tidy.svyglm: svyglm

tidy.svyolr: svyolr

tidy.systemfit: systemfit

tidy.table: table

tidy.ts: ts

tidy.TukeyHSD: TukeyHSD

tidy.varest: varest

tidy.zoo: zoo

vector_tidiers: character, logical, numeric

broom.mixed

brms_tidiers: brmsfit

gamlss_tidiers: gamlss

glmmadmb_tidiers: glmmadmb

glmmTMB_tidiers: glmmTMB

lme4_tidiers: merMod, rlmerMod

mcmc_tidiers: mcmc, mcmc.list, MCMCglmm, rjags, stanfit

nlme_tidiers: gls, lme

rstanarm_tidiers: stanreg

tidy.TMB: TMB

See also

Examples

set.seed(123)

# linear mixed effects model
broomExtra::grouped_tidy(
  data = dplyr::mutate(MASS::Aids2, interval = death - diag),
  grouping.vars = sex,
  ..f = lme4::lmer,
  formula = interval ~ age + (1 | status),
  control = lme4::lmerControl(optimizer = "bobyqa"),
  tidy.args = list(conf.int = TRUE, conf.level = 0.99)
)

#> # A tibble: 8 x 9
#>   sex   effect  group  term      estimate std.error statistic conf.low conf.high
#>   <fct> <chr>   <chr>  <chr>        <dbl>     <dbl>     <dbl>    <dbl>     <dbl>
#> 1 F     fixed   NA     (Interce…   623.     161.         3.88   209.     1037.  
#> 2 F     fixed   NA     age          -4.34     2.61      -1.66   -11.1       2.38
#> 3 F     ran_pa… status sd__(Int…   169.      NA         NA       NA        NA   
#> 4 F     ran_pa… Resid… sd__Obse…   415.      NA         NA       NA        NA   
#> 5 M     fixed   NA     (Interce…   553.      62.5        8.84   392.      714.  
#> 6 M     fixed   NA     age          -3.60     0.696     -5.17    -5.39     -1.80
#> 7 M     ran_pa… status sd__(Int…    79.8     NA         NA       NA        NA   
#> 8 M     ran_pa… Resid… sd__Obse…   355.      NA         NA       NA        NA