Tidy output from grouped analysis of any function that has data argument in its function call.

Tidy output from grouped analysis of any function that has data argument in its function call.

grouped_tidy(data, grouping.vars, ..f, ..., tidy.args = list())

Arguments

data
Dataframe (or tibble) from which variables are to be taken.

grouping.vars
Grouping variables.

..f
A function, or function name as a string.

...
<dynamic> Arguments for .fn.

tidy.args
A list of arguments to be used in the relevant S3 method.

Value

A tibble::tibble() with information about model components.

Note

For available methods, see- https://indrajeetpatil.github.io/broomExtra/articles/available_methods.html

Methods

See the following help topics for more details about individual methods:broom

durbinWatsonTest_tidiers: durbinWatsonTest

list_tidiers: list

metafor_tidiers: rma

mgcv_tidy_gam: gam

null_tidiers: NULL

sp_tidiers: Line, Lines, Polygon, Polygons, SpatialLinesDataFrame, SpatialPolygons, SpatialPolygonsDataFrame

tidy.aareg: aareg

tidy.acf: acf

tidy.anova: anova

tidy.aov: aov

tidy.aovlist: aovlist

tidy.Arima: Arima

tidy.betareg: betareg

tidy.biglm: biglm

tidy.binDesign: binDesign

tidy.binWidth: binWidth

tidy.boot: boot

tidy.btergm: btergm

tidy.cch: cch

tidy.cld: cld

tidy.clm: clm

tidy.clmm: clmm

tidy.coeftest: coeftest

tidy.confint.glht: confint.glht

tidy.confusionMatrix: confusionMatrix

tidy.coxph: coxph

tidy.cv.glmnet: cv.glmnet

tidy.density: density

tidy.dist: dist

tidy.drc: drc

tidy.emmGrid: emmGrid

tidy.epi.2by2: epi.2by2

tidy.ergm: ergm

tidy.factanal: factanal

tidy.felm: felm

tidy.fitdistr: fitdistr

tidy.ftable: ftable

tidy.Gam: Gam

tidy.garch: garch

tidy.geeglm: geeglm

tidy.glht: glht

tidy.glm: glm

tidy.glmnet: glmnet

tidy.glmRob: glmRob

tidy.gmm: gmm

tidy.htest: htest

tidy.ivreg: ivreg

tidy.kappa: kappa

tidy.kde: kde

tidy.Kendall: Kendall

tidy.kmeans: kmeans

tidy.lavaan: lavaan

tidy.lm.beta: lm.beta

tidy.lm: lm

tidy.lmodel2: lmodel2

tidy.lmRob: lmRob

tidy.lsmobj: lsmobj

tidy.manova: manova

tidy.map: map

tidy.Mclust: Mclust

tidy.mediate: mediate

tidy.mjoint: mjoint

tidy.mle2: mle2

tidy.mlm: mlm

tidy.muhaz: muhaz

tidy.multinom: multinom

tidy.nlrq: nlrq

tidy.nls: nls

tidy.orcutt: orcutt

tidy.pairwise.htest: pairwise.htest

tidy.pam: pam

tidy.plm: plm

tidy.poLCA: poLCA

tidy.polr: polr

tidy.power.htest: power.htest

tidy.prcomp: prcomp

tidy.pyears: pyears

tidy.rcorr: rcorr

tidy.ref.grid: ref.grid

tidy.regsubsets: regsubsets

tidy.ridgelm: ridgelm

tidy.rlm: rlm

tidy.robustbase.glmrob: glmrob

tidy.robustbase.lmrob: lmrob

tidy.roc: roc

tidy.rq: rq

tidy.rqs: rqs

tidy.spec: spec

tidy.speedglm: speedglm

tidy.speedlm: speedlm

tidy.summary.glht: summary.glht

tidy.summary.manova: summary.manova

tidy.summary_emm: summary_emm

tidy.survdiff: survdiff

tidy.survexp: survexp

tidy.survfit: survfit

tidy.survreg: survreg

tidy.svyglm: svyglm

tidy.svyolr: svyolr

tidy.systemfit: systemfit

tidy.table: table

tidy.ts: ts

tidy.TukeyHSD: TukeyHSD

tidy.zoo: zoo

broom.mixed

brms_tidiers: brmsfit

gamlss_tidiers: gamlss

glmmadmb_tidiers: glmmadmb

glmmTMB_tidiers: glmmTMB

lme4_tidiers: merMod, rlmerMod

mcmc_tidiers: mcmc, mcmc.list, MCMCglmm, rjags, stanfit

nlme_tidiers: gls, lme

rstanarm_tidiers: stanreg

tidy.TMB: TMB

See also

Examples

set.seed(123)
# to speed up computation, let's use only 50% of the data

# linear model
broomExtra::grouped_tidy(
  data = dplyr::sample_frac(tbl = ggplot2::diamonds, size = 0.5),
  grouping.vars = c(cut, color),
  formula = price ~ carat - 1,
  ..f = stats::lm,
  na.action = na.omit,
  tidy.args = list(quick = TRUE)
)

#> # A tibble: 35 x 7
#>    cut   color term  estimate std.error statistic   p.value
#>    <ord> <ord> <chr>    <dbl>     <dbl>     <dbl>     <dbl>
#>  1 Fair  D     carat    5246.     207.       25.3 4.45e- 41
#>  2 Fair  E     carat    4202.     158.       26.6 3.52e- 47
#>  3 Fair  F     carat    4877.     149.       32.7 1.68e- 71
#>  4 Fair  G     carat    4538.     152.       29.8 1.03e- 66
#>  5 Fair  H     carat    4620.     146.       31.6 7.68e- 66
#>  6 Fair  I     carat    3969.     136.       29.2 4.86e- 44
#>  7 Fair  J     carat    4024.     197.       20.4 4.80e- 27
#>  8 Good  D     carat    5207.     115.       45.4 2.66e-145
#>  9 Good  E     carat    5102.      91.9      55.5 2.50e-206
#> 10 Good  F     carat    5151.      92.4      55.8 1.76e-204
#> # ... with 25 more rows

# linear mixed effects model
broomExtra::grouped_tidy(
  data = dplyr::sample_frac(tbl = ggplot2::diamonds, size = 0.5),
  grouping.vars = cut,
  ..f = lme4::lmer,
  formula = price ~ carat + (carat | color) - 1,
  control = lme4::lmerControl(optimizer = "bobyqa"),
  tidy.args = list(conf.int = TRUE, conf.level = 0.99)
)

#> # A tibble: 25 x 9
#>    cut   effect   group    term                   estimate std.error statistic
#>    <ord> <chr>    <chr>    <chr>                     <dbl>     <dbl>     <dbl>
#>  1 Fair  fixed    NA       carat                  3800.         228.      16.7
#>  2 Fair  ran_pars color    sd__(Intercept)        2158.          NA       NA  
#>  3 Fair  ran_pars color    cor__(Intercept).carat   -0.975       NA       NA  
#>  4 Fair  ran_pars color    sd__carat              2545.          NA       NA  
#>  5 Fair  ran_pars Residual sd__Observation        1830.          NA       NA  
#>  6 Good  fixed    NA       carat                  9217.         105.      87.6
#>  7 Good  ran_pars color    sd__(Intercept)        2686.          NA       NA  
#>  8 Good  ran_pars color    cor__(Intercept).carat    0.998       NA       NA  
#>  9 Good  ran_pars color    sd__carat              1609.          NA       NA  
#> 10 Good  ran_pars Residual sd__Observation        1373.          NA       NA  
#>    conf.low conf.high
#>       <dbl>     <dbl>
#>  1    3212.     4387.
#>  2      NA        NA 
#>  3      NA        NA 
#>  4      NA        NA 
#>  5      NA        NA 
#>  6    8946.     9488.
#>  7      NA        NA 
#>  8      NA        NA 
#>  9      NA        NA 
#> 10      NA        NA 
#> # ... with 15 more rows