Retrieve tidy dataframe if it exists.

Checks if a tidy method exits for a given object, either in broom or in broom.mixed. If it does, it turn an object into a tidy tibble, if not, return a NULL. In case of data frames, a tibble data frame is returned. In this case, you can try the broomExtra::tidy_parameters function.

tidy(x, ...)

Arguments

x
An object to be converted into a tidy tibble::tibble().

...
Additional arguments to tidying method.

Value

A tibble::tibble() with information about model components.

Note

For available methods, see- https://indrajeetpatil.github.io/broomExtra/articles/available_methods.html

Methods

See the following help topics for more details about individual methods:broom

durbinWatsonTest_tidiers: durbinWatsonTest

list_tidiers: list

metafor_tidiers: rma

mgcv_tidy_gam: gam

null_tidiers: NULL

sp_tidiers: Line, Lines, Polygon, Polygons, SpatialLinesDataFrame, SpatialPolygons, SpatialPolygonsDataFrame

tidy.aareg: aareg

tidy.acf: acf

tidy.anova: anova

tidy.aov: aov

tidy.aovlist: aovlist

tidy.Arima: Arima

tidy.betareg: betareg

tidy.biglm: biglm

tidy.binDesign: binDesign

tidy.binWidth: binWidth

tidy.boot: boot

tidy.btergm: btergm

tidy.cch: cch

tidy.cld: cld

tidy.clm: clm

tidy.clmm: clmm

tidy.coeftest: coeftest

tidy.confint.glht: confint.glht

tidy.confusionMatrix: confusionMatrix

tidy.coxph: coxph

tidy.cv.glmnet: cv.glmnet

tidy.density: density

tidy.dist: dist

tidy.drc: drc

tidy.emmGrid: emmGrid

tidy.epi.2by2: epi.2by2

tidy.ergm: ergm

tidy.factanal: factanal

tidy.felm: felm

tidy.fitdistr: fitdistr

tidy.ftable: ftable

tidy.Gam: Gam

tidy.garch: garch

tidy.geeglm: geeglm

tidy.glht: glht

tidy.glm: glm

tidy.glmnet: glmnet

tidy.glmRob: glmRob

tidy.gmm: gmm

tidy.htest: htest

tidy.ivreg: ivreg

tidy.kappa: kappa

tidy.kde: kde

tidy.Kendall: Kendall

tidy.kmeans: kmeans

tidy.lavaan: lavaan

tidy.lm.beta: lm.beta

tidy.lm: lm

tidy.lmodel2: lmodel2

tidy.lmRob: lmRob

tidy.lsmobj: lsmobj

tidy.manova: manova

tidy.map: map

tidy.Mclust: Mclust

tidy.mediate: mediate

tidy.mjoint: mjoint

tidy.mle2: mle2

tidy.mlm: mlm

tidy.muhaz: muhaz

tidy.multinom: multinom

tidy.nlrq: nlrq

tidy.nls: nls

tidy.orcutt: orcutt

tidy.pairwise.htest: pairwise.htest

tidy.pam: pam

tidy.plm: plm

tidy.poLCA: poLCA

tidy.polr: polr

tidy.power.htest: power.htest

tidy.prcomp: prcomp

tidy.pyears: pyears

tidy.rcorr: rcorr

tidy.ref.grid: ref.grid

tidy.regsubsets: regsubsets

tidy.ridgelm: ridgelm

tidy.rlm: rlm

tidy.robustbase.glmrob: glmrob

tidy.robustbase.lmrob: lmrob

tidy.roc: roc

tidy.rq: rq

tidy.rqs: rqs

tidy.spec: spec

tidy.speedglm: speedglm

tidy.speedlm: speedlm

tidy.summary.glht: summary.glht

tidy.summary.manova: summary.manova

tidy.summary_emm: summary_emm

tidy.survdiff: survdiff

tidy.survexp: survexp

tidy.survfit: survfit

tidy.survreg: survreg

tidy.svyglm: svyglm

tidy.svyolr: svyolr

tidy.systemfit: systemfit

tidy.table: table

tidy.ts: ts

tidy.TukeyHSD: TukeyHSD

tidy.zoo: zoo

broom.mixed

brms_tidiers: brmsfit

gamlss_tidiers: gamlss

glmmadmb_tidiers: glmmadmb

glmmTMB_tidiers: glmmTMB

lme4_tidiers: merMod, rlmerMod

mcmc_tidiers: mcmc, mcmc.list, MCMCglmm, rjags, stanfit

nlme_tidiers: gls, lme

rstanarm_tidiers: stanreg

tidy.TMB: TMB

See also

Examples

set.seed(123)
library(lme4)

# mixed-effects models (`broom.mixed` will be used)
lmm.mod <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
broomExtra::tidy(x = lmm.mod, effects = "fixed", exponentiate = TRUE)

#> # A tibble: 2 x 5
#>   effect term        estimate std.error statistic
#>   <chr>  <chr>          <dbl>     <dbl>     <dbl>
#> 1 fixed  (Intercept) 1.53e109  1.04e110     36.8 
#> 2 fixed  Days        3.51e  4  5.43e  4      6.77

# linear model (`broom` will be used)
lm.mod <- lm(Reaction ~ Days, sleepstudy)
broomExtra::tidy(x = lm.mod, conf.int = TRUE)

#> # A tibble: 2 x 7
#>   term        estimate std.error statistic  p.value conf.low conf.high
#> * <chr>          <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl>
#> 1 (Intercept)    251.       6.61     38.0  2.16e-87   238.       264. 
#> 2 Days            10.5      1.24      8.45 9.89e-15     8.02      12.9

# unsupported object (the function will return `NULL` in such cases)
broomExtra::tidy(list(1, c("x", "y")))

#> NULL